backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
aiosqlite==0.20.0
alembic==1.13.3
psycopg2-binary==2.9.9
celery==5.4.0
redis==5.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12
minio==7.2.9
openai==1.51.0
ffmpeg-python==0.2.0
scenedetect[opencv]==0.6.4
librosa==0.10.2
httpx==0.27.2
pytest==8.3.3
pytest-asyncio==0.24.0
anyio==4.4.0
